Fix SFN.tag_description.keyword problem
Certain malware tags (like NJRat) were not populating in the visualizations
because they had special characters - NJRat had parenthesis () - that
would cause it to not display.  This required adding the mapping to the
DB so it would save it with the keyword info as well as string and that
allows it to display in the tables for the visualizations.